refactor to use custom hook
